Refactor HotReloadAgent in to a separate type
This allows for easier sharing of the agent code with Blazor WebAssembly's implementation.
