+++ src/BuiltInTools/dotnet-watch/HotReload/NamedPipeContract.cs
