+++ src/Layout/redist/targets/OverlaySdkOnLKG.targets
@@ -45,7 +45,7 @@
     <ItemGroup>
       <DotNetWatchOverlay Include="$(ArtifactsDir)bin\dotnet-watch\$(Configuration)\$(SdkTargetFramework)\**" />
       <DotNetWatchOverlay Include="$(ArtifactsDir)bin\Microsoft.AspNetCore.Watch.BrowserRefresh\$(Configuration)\netcoreapp3.1\*.dll" TargetDir="middleware" />
-      <DotNetWatchOverlay Include="$(ArtifactsDir)bin\Microsoft.Extensions.AspNetCoreDeltaApplier\$(Configuration)\netcoreapp3.1\*.dll" TargetDir="hotreload" />
+      <DotNetWatchOverlay Include="$(ArtifactsDir)bin\Microsoft.Extensions.DotNetDeltaApplier\$(Configuration)\net6.0\*.dll" TargetDir="hotreload" />
       <DotNetWatchOverlay Include="$(ArtifactsDir)bin\DotNetWatchTasks\$(Configuration)\netstandard2.0\DotNetWatchTasks.dll" />
     </ItemGroup>
